Support libcurl pause/unpause.
Libcurl has the ability to 'pause' a connection temporarily
and then resume it when more data is available (etc.). This
patch makes it possible to use the pause/unpause features of
libcurl via pycurl. It adds a small number of constants, a
new method on Curl objects, and some documentation.

http://sourceforge.net/p/pycurl/patches/11/