Add suport for CURLINFO_CERTINFO and CURLOPT_CERTINFO.
This patch adds support for CURLINFO_CERTINFO (and CURLOPT_CERTINFO),
as well as documentation for them and a couple of minor documentation fixes.
These options appeared in libcurl 7.19.1.

The format of the returned information was chosen to be reasonably close
to the underlying libcurl data structure, while also allowing a
Python programmer to pass the cerificate info to dict() in order to get
a more convenient representation.

https://sourceforge.net/p/pycurl/patches/14/