Only initialize gcrypt if we are using an older gnutls that needs this.
This is necessary to support newer gnutls linked to libnettle instead
of libgcrypt.

If curl is using a recent version of gnutls we do not need to initialize
libgcrypt threading: gnutls does that for us. And because recent versions
of gnutls may be using nettle instead of gcrypt it is important that we
don't: those functions are not necessarily defined (and linking to
libgcrypt explicitly to get them is stupid, as they're not used). The
modification is based on instructions from the gnutls NEWS file.

http://sourceforge.net/p/pycurl/patches/15/