Mark NSS as supported.
Mark NSS as supported, as it does not require the application to
initialize threading.

If curl is using nss we do not seem to need to initialize threading
explicitly: the only threading-related bit I see in the nss
headers/documentation has to do with simultaneous non-nss use of
pkcs11 modules, nss itself seems to be thread-safe by default.
So silence the build warning.

http://sourceforge.net/p/pycurl/patches/15/