vsftpd.conf: add background=no to allow for proper shutdown
When the "background" directive in the vsftpd.conf configuration file is
set to "YES", the vsftpd startup script forks, creating a child process
(the vsftpd daemon) which immediately sends the SIGUSR1 signal to its
parent process, which exits upon receiving it.  The teardown routine in
procmgr.py would then send the SIGTERM signal to a process that does not
exist anymore.