Fix memory leaks in curl's POSTFIELDS setopt for unicode objects.
Fix a memory leak in Python 2 when passing a unicode object
to a curl object's POSTFIELDS setopt where the str object
created from the passed in unicode object would be leaked on
the successful path.

Fix memory leak in the curl object's POSTFIELDS setopt where
if the curl_easy_setopt CURLOPT_POSTFIELDSIZE or
CURLOPT_POSTFIELDSIZE_LARGE failed the str object created
from the passed in unicode object would be leaked.