end trashcan sections after calling free
The Py_TRASHCAN_SAFE_* macros create a lexical block that may or may
not be executed depending on the stack depth. If it's not executed,
the deallocator is called again when the stack is not so deeply
nested. The current curl deallocators have the problem that if the
block is not executed, the object is still freed because
Py_TYPE(self)->tp_free is called outside of the trashcan block. This
can lead to double-freeing.