Add consistent names for newer Curl version constants
Unfortunately, in commit 05bc44d, a bunch of Curl version constants were
introduced with a leading "CURL_" in their names, which is inconsistent
with the rest of the version constants.  This code has been out in the
wild for two years, so it seems that we can't just correct this mistake.
Instead, just introduce the correct names while leaving the old ones.
Perhaps in the future the old ones can be deprecated/removed.

Also, add unit tests for the version constants.