macOS CI: build using Homebrew's libcurl instead
It has become increasingly difficult to use the version of libcurl
shipped by Apple, so switch to building with Homebrew's version instead
(but still use SecureTransport).  The main difficulty is that Apple
doesn't ship the .dylib that matches the SDK version (which includes
curl.h), nor does it ship the curl.h that matches the installed .dylib.
So it has become a mess of mismatched versions.