Fix curl_multi_info_read flow that loses messages
Fixes an issue where (if num_results is less than the number of
messages in the queue) a message can be pulled out of the queue, then
the flow breaks out of the loop before the message is added to one of
the returned lists.  The user therefore never sees the message.